Move constraint to_description tests to proper location
- Linear constraint tests → tests/bofire/data_models/constraints/test_linear.py
- NChooseK + Product tests → tests/bofire/data_models/constraints/test_to_description.py
- Remove from tests/bofire/data_models/domain/test_constraints.py
