+++ tests/bofire/data_models/domain/test_constraints.py
@@ -150,72 +150,3 @@ def test_constraints_jacobian(constraints, num_candidates):
                 if not hasattr(col, "__iter__"):
                     res[j] = pd.Series(np.repeat(col, candidates.shape[0]))
             assert np.allclose(returned[i], pd.DataFrame(res).transpose())
-
-
-def test_linear_equality_constraint_to_description():
-    c = LinearEqualityConstraint(
-        features=["x1", "x2"], coefficients=[1.0, 2.0], rhs=5.0
-    )
-    desc = c.to_description()
-    assert "1.0*x1" in desc
-    assert "2.0*x2" in desc
-    assert "= 5.0" in desc
-
-
-def test_linear_inequality_constraint_to_description():
-    c = LinearInequalityConstraint(
-        features=["x1", "x2"], coefficients=[1.0, 2.0], rhs=5.0
-    )
-    assert "<= 5.0" in c.to_description()
-
-
-def test_linear_constraint_to_description_with_context():
-    c = LinearInequalityConstraint(
-        features=["x1", "x2"],
-        coefficients=[1.0, 2.0],
-        rhs=5.0,
-        context="Safety limit",
-    )
-    assert "Safety limit" in c.to_description()
-
-
-def test_nchoosek_constraint_to_description():
-    c = NChooseKConstraint(
-        features=["x1", "x2", "x3"],
-        min_count=1,
-        max_count=2,
-        none_also_valid=False,
-    )
-    desc = c.to_description()
-    assert "Choose 1-2" in desc
-    assert "x1" in desc
-
-
-def test_nchoosek_constraint_to_description_none_valid():
-    c = NChooseKConstraint(
-        features=["x1", "x2"],
-        min_count=1,
-        max_count=2,
-        none_also_valid=True,
-    )
-    assert "or none" in c.to_description()
-
-
-def test_product_equality_constraint_to_description():
-    from bofire.data_models.constraints.api import ProductEqualityConstraint
-
-    c = ProductEqualityConstraint(
-        features=["x1", "x2"], exponents=[2, 3], rhs=1.0, sign=1
-    )
-    desc = c.to_description()
-    assert "x1^2" in desc
-    assert "= 1.0" in desc
-
-
-def test_product_inequality_constraint_to_description():
-    from bofire.data_models.constraints.api import ProductInequalityConstraint
-
-    c = ProductInequalityConstraint(
-        features=["x1", "x2"], exponents=[2, 3], rhs=1.0, sign=1
-    )
-    assert "<= 1.0" in c.to_description()
